Add files via upload
It looks like I do not have enough knowledge of NativeScript to finish the app. I will have to stop here.

I have drawn the bord and set up tap habdler funtions.
However, to modify the TicTcToe board, I need to get view for GridLayout. It seems I need to create page view, not component to use getViewById().getChildAt(index) to access each cell.
When player tap a cell, the function will receive the cell index as argument, and the event handler will be able to update the test on the cell. So is the AI.

Move code to right loction.